+++ cmd/fofa/cmd/pipeline.go
@@ -1,24 +1,20 @@
 package cmd
 
 import (
-	"bytes"
 	"errors"
 	"fmt"
 	"github.com/lubyruffy/gofofa/pkg/funcs"
-	"github.com/lubyruffy/gofofa/pkg/outformats"
 	"github.com/lubyruffy/gofofa/pkg/pipeast"
 	"github.com/lubyruffy/gofofa/pkg/piperunner"
-	"github.com/mitchellh/mapstructure"
 	"github.com/urfave/cli/v2"
 	"io/ioutil"
 	"os"
-	"strings"
-	"text/template"
 )
 
 var (
 	pipelineFile    string
 	pipelineTaskOut string // 导出任务列表文件
+	listWorkflows   bool
 )
 
 // pipeline subcommand
@@ -39,6 +35,12 @@ var pipelineCmd = &cli.Command{
 			Usage:       "output pipeline tasks",
 			Destination: &pipelineTaskOut,
 		},
+		&cli.BoolFlag{
+			Name:        "list",
+			Aliases:     []string{"l"},
+			Usage:       "list support workflows",
+			Destination: &listWorkflows,
+		},
 	},
 	Action: pipelineAction,
 }
@@ -48,8 +50,10 @@ var pipelineCmd = &cli.Command{
 // 也就是说注册一个pipeline可以支持的命令，需要：一）注册底层函数；二）注册pipeline的函数到底层函数调用的代码转换器
 func pipelineAction(ctx *cli.Context) error {
 
-	funcs.Load()
-	piperunner.RegisterWorkflow("fofa", fofaHook, "FetchFofa", fetchFofa)
+	if listWorkflows {
+		fmt.Println(funcs.SupportWorkflows())
+		return nil
+	}
 
 	// valid same config
 	var pipelineContent string
@@ -68,6 +72,7 @@ func pipelineAction(ctx *cli.Context) error {
 	}
 
 	pr := piperunner.New(pipelineContent)
+	pr.FofaCli = fofaCli
 	err := pr.Run()
 	if err != nil {
 		return err
@@ -89,73 +94,4 @@ func pipelineAction(ctx *cli.Context) error {
 	}
 
 	return nil
-}
-
-type fetchFofaParams struct {
-	Query  string
-	Size   int
-	Fields string
-}
-
-func fetchFofa(p *piperunner.PipeRunner, params map[string]interface{}) string {
-	var err error
-	var options fetchFofaParams
-	if err = mapstructure.Decode(params, &options); err != nil {
-		panic(err)
-	}
-
-	if len(options.Query) == 0 {
-		panic(errors.New("fofa query cannot be empty"))
-	}
-	if len(options.Fields) == 0 {
-		panic(errors.New("fofa fields cannot be empty"))
-	}
-
-	fields := strings.Split(options.Fields, ",")
-
-	var res [][]string
-	res, err = fofaCli.HostSearch(options.Query, options.Size, fields)
-	if err != nil {
-		panic(err)
-	}
-
-	return piperunner.WriteTempJSONFile(func(f *os.File) {
-		w := outformats.NewJSONWriter(f, fields)
-		if err = w.WriteAll(res); err != nil {
-			panic(err)
-		}
-	})
-}
-
-func fofaHook(fi *pipeast.FuncInfo) string {
-	tmpl, err := template.New("fofa").Parse(`FetchFofa(GetRunner(), map[string]interface{} {
-    "query": {{ .Query }},
-    "size": {{ .Size }},
-    "fields": {{ .Fields }},
-})`)
-	if err != nil {
-		panic(err)
-	}
-	var size int64 = 10
-	fields := "`host,title`"
-	if len(fi.Params) > 1 {
-		fields = fi.Params[1].String()
-	}
-	if len(fi.Params) > 2 {
-		size = fi.Params[2].Int64()
-	}
-	var tpl bytes.Buffer
-	err = tmpl.Execute(&tpl, struct {
-		Query  string
-		Size   int64
-		Fields string
-	}{
-		Query:  fi.Params[0].String(),
-		Fields: fields,
-		Size:   size,
-	})
-	if err != nil {
-		panic(err)
-	}
-	return tpl.String()
 }