+++ pkg/funcs/funcs.go
@@ -1,5 +1,8 @@
 package funcs
 
-// Load 手动加载，否则init不执行
-func Load() {
+import "github.com/lubyruffy/gofofa/pkg/pipeast"
+
+// SupportWorkflows 手动加载，否则init不执行
+func SupportWorkflows() []string {
+	return pipeast.SupportWorkflows()
 }